requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
gradio==4.7.1
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
pydantic==2.5.2
boto3==1.29.6
aiohttp==3.9.1
pytest==7.4.3
fugashi==1.2.1
ipadic==1.0.0
unidic-lite==1.0.8 